cowebsite navigation sound player extraction
