+++ front/src/Api/iframe/Player.ts
@@ -1,0 +1,22 @@
+import { IframeApiContribution, sendToWorkadventure } from './IframeApiContribution';
+
+class WorkadventureNavigationCommands extends IframeApiContribution<WorkadventureNavigationCommands> {
+
+    readonly subObjectIdentifier = "player"
+
+    readonly addMethodsAtRoot = true
+    callbacks = []
+
+    disablePlayerControls(): void {
+        sendToWorkadventure({ 'type': 'disablePlayerControls', data: null });
+    }
+
+    restorePlayerControls(): void {
+        sendToWorkadventure({ 'type': 'restorePlayerControls', data: null });
+    }
+
+
+}
+
+
+export default new WorkadventureNavigationCommands();
